Improve directory tree
