+++ spotdl/encode/__init__.py
@@ -1,0 +1,1 @@
+from spotdl.encode.encode_base import EncoderBase
