+++ spotdl/encode/encoders/__init__.py
@@ -1,0 +1,4 @@
+from spotdl.encode.encoders.ffmpeg import EncoderFFmpeg
+from spotdl.encode.encoders.avconv import EncoderAvconv
+
+EncodeClasses = (EncoderFFmpeg, EncoderAvconv)
