--- spotdl/encoders/__init__.py
@@ -1,3 +1,0 @@
-from spotdl.encoders.base import EncoderBase
-from spotdl.encoders.ffmpeg import EncoderFFmpeg
-from spotdl.encoders.avconv import EncoderAvconv
